Status Bar Auto Hide for tablets (1/2) Also compatible apps to auto hide if full screen. IE youtube, gallery, ect.
Change-Id: I1334ec5bfeab269666ea725e06ba631c6698a8b0
[Stevespear426]

Sliding NavBar Widgets Control
ROMControl will allow an unlimited number of widgets
and will attempt to keep the widget DB clean by
De-allocating ID's when re-assigned.
If NavBar is reset, RC will ask SystemUI to
Deallocate each of the ID's in the prefs.
There is now a 'Reset Nav Widgets' option in
the action menu that will reset just the widgets
without affecting the rest of the NavBar.

Signed-off-by: Mike Wilson <mwilson@acesappraisal.com>
Change-Id: I87571b1bd5eb056ad136a2ba75715a4b86301ff8
[Zaphod]